requirements.txt
cffi==1.9.1
cryptography==1.7.1
idna==2.2
paramiko==2.1.1
pyasn1==0.1.9
pycparser==2.17
scapy-python3==0.20
six==1.10.0
wheel==0.24.0
